+++ extensions/reclaim-ai/CHANGELOG.md
@@ -1,5 +1,11 @@
 # reclaim Changelog
 
+## [Update] - 2024-07-16
+- Add actions for Smart Meetings & Habits 2.0
+- Parity between Raycast actions and Omnibar Actions
+- Fix "reschedule event" to work for Tasks and use new endpoint
+- TODO: update tasks/smart habits start/stop to respect auto reschedule
+
 ## [Update] - 2024-05-10
 - Cache the user object for 30 minute for a small performance benefit.